pyproject.toml
[project]
name = "dsozluk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepl>=1.22.0",
    "flask>=3.1.1",
    "openai>=1.93.1",
    "python-dotenv>=1.1.1",
    "gunicorn",
    "instructor>=1.10.0",
]
